Renamed ServerTransportTestBase to TransportTestBase. Refactored existing tests accordingly.
